Add a general way to introduce trusted impurities in builders.
If a derivation sets __useImpurityHelper to true, then the
NIX_IMPURITY_SOCKET env var will be set to a file descriptor that is a
unix domain socket. To call a trusted impure command, the builder:

1. Writes a byte to the main socket
2. Reads a file descriptor (that corresponds to a new socket) from the
main socket
3. Writes the name of the program as a null-terminated character string
to the new socket
4. Writes the number of arguments to the program as an int (as defined
by the local C++ ABI) to the new socket
5. Consecutively writes a set of null-terminated arguments to the new
socket (as many as specified in 4)
6. Recieves a file descriptor corresponding to the write side of stdin
of the trusted impure command from the new socket
7. Recieves a file descriptor corresponding to the read side of stdout
of the trusted impure command from the new socket
8. Recieves a file descriptor corresponding to the read side of stderr
of the trusted impure command from the new socket
9. If the trusted impure command exits normally, reads a char
corresponding to its exit status and then an EOF from the new socket
10. If the trusted impure command fails due to a problem before execve,
reads a single byte representing 0 from the new socket then reads the
errno (represented as an int according to the local C++ ABI) from the
socket, followed by an EOF
11. If the trusted impure command fails due to a signal, reads a single
byte representing a 1 then reads the signal number (represented as an
int according to the local C++ ABI) from the socket, follwed by an EOF.

The impurity-helper, upon recieving a request, will look for a command
with the given name in the directory specified by the
impure-commands-dir nix.conf option (defaults to /var/empty) and
executes it (outside the chroot!) with argv[0] set to the binary path,
argv[1] set to the name of the .drv file of the build, argv[2] set to
the temporary directory of the build, argv[3] set to the chroot root
of the build (the empty stringif it is not in chroot), and the rest
of the arguments set to those passed by the builder over the socket.

The purpose of this facility is to safely enable tools that break the
rules imposed on nix builders but don't actually violate the invariants
those rules are meant to guarantee. An example might be ccache: For
ccache to work, there must be a global database that every builder can
acccess, but if this were implemented naievely a builder could
trojan-horse another user by putting invalid entries in the cache. With
this mechanism, a single trusted script would be the only one that had
write access to the cache.