Add a new struct Derivation
Outputs are maps from name to hash info, inputs are paths. There's
no need to precalculate the output hashes, and inputs will all affect
the result hash the same regardless of their source, i.e. if a source,
a derivation, and a fixed-output derivation create the same output path
they're treated the same in inputs.