Unify handling of foo.drvPath and storePath /nix/store/...-foo.drv
Previously, derivation files added to the context via
builtins.storePath were treated *almost* like derivation files
accessed directly through the drvPath attribute, except that the
inputs to the drv file weren't recursively searched in the former
case. This fixes that discrepancy.

This is technically a backwards incompatibility, but storePath on
derivation files is never used in nixpkgs and extremely rare
elsewhere, and if the context behavior is desired this is almost
certainly the intended semantics.